Moved buildsettings for scalikejdbc-play-plugin/test/zentask to build/Build.scala.
scalikejdbc-play-plugin was intended to be a sample application of scalikejdbc-play-plugin
but it is convenience for scalikejdbc-play-plugin's test to use this project.
This fix makes scalikejdbc-play-plugin depend on and aggregate other sub projects.
I don't need to do 'publish-local' every time I change the code in play-plugin for manual test.

Two points in my mind (may be trivial).
 - Works only with > Scala 2.10 because it depends on scalikejdbc-interpolation.
 - Play 2.1 depends on Scala 2.10.0 but some parts of interpolation is based on macro's features that seem to be incompatible between 2.10.0 and 2.10.1.