Configurable warm up time for new connection pools
Make the time to wait after adding a connection pool to scalikejdbc
configurable through a new `poolWarmUpTime`-setting in the database
config.

I decided to make this setting on the connection pool level instead of
in `GlobalSettings` because it might very well be something that makes
sense to configure differently based on the connection pool being
used.