Fix SQLInterpolation of java.time.LocalTime to deal with DST.
Caluculating zoned `java.sql.Time` using `java.time.OffsetTime` ignores
DST. This commit fix the conversion of `java.time.LocalTime` to `java.sql.Time` to deal with DST.
We set the date components of the instance of java.time.LocalTime to 1970-01-01 on converting it to `java.sql.Time` following the javadoc.

The date components should be set to the "zero epoch" value of January 1, 1970 and should not be accessed.
https://docs.oracle.com/javase/8/docs/api/java/sql/Time.html